created a basic css to improve the look of the page , i have added a back button to the html
