+++ src/html/benefit.html
@@ -1,20 +1,23 @@
-<?php require_once 'config.php'; ?>
-
 <!DOCTYPE html>
 <html lang="en">
 <head>
   <meta charset="UTF-8">
   <title>Benefit Form</title>
+  <link rel="stylesheet" href="../css/style.css">
 </head>
 <body>
   <h1>Benefit Form</h1>
   <form action="../php/benefit.php" method="post">
    
-    <label for="benefit_name">Benefit Name:</label>
+    <label for="benefit_name">Benefit Name</label>
     <input type="text" name="benefit_name" id="benefit_name" maxlength="50" required><br><br>
 
     <input type="submit" name="submit_benefit" value="Submit">
+    <a type="submit" href="../index.php">
+      <button type="button">Back</button>
+    </a>
   </form>
+  
 </body>
 </html>
