+++ src/html/location.html
@@ -2,19 +2,23 @@
 <html lang="en">
 <head>
   <meta charset="UTF-8">
-  <title>Forms Index</title>
+  <title>Location Form</title>
+  <link rel="stylesheet" href="../css/style.css">
 </head>
 <body>  
-  <h2>Location Form</h2>
+  <h1>Location Form</h1>
   <form action="../php/location.php" method="POST">
-    <label for="loc_name">Location Name:</label>
+    <label for="loc_name">Location Name</label>
     <input type="text" name="loc_name" id="loc_name" required>
 
-    <label for="address">Address:</label>
+    <label for="address">Address</label>
     <input type="text" name="address" id="address" required>
 
-    <label for="country">Country:</label>
+    <label for="country">Country</label>
     <input type="text" name="country" id="country" required>
 
     <input type="submit" name="submit_location" value="Submit">
+    <a type="submit" href="../index.php">
+      <button type="button">Back</button>
+    </a>
   </form>
